backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pandas==2.2.3
numpy==2.2.1
python-multipart==0.0.18
aiofiles==24.1.0
python-dateutil==2.9.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.0
